Made the update rate of weather configurable.
Fixed again a problem: The weather was updated initially from different
threads in parallel. This caused a higher load and request rate of the
webservice API. When developing this can (still) cause that server deny
the weather or geocoder access from this specific IP.

If it fails we simply ignore it now. (Needs to be checked if this is
valid behaviour or if we need to visualize this better)

Moved the update check to the abstract base class. So just one place to
maintain.

The Wunderground provider had two magic lines:
  Idle.updateIdle(context, true);
  MetaWatchService.notifyClients();
Moved them to base class cause they look like we should call them in the
other providers too?